Cleanup IDocumentEditorView duplicate definitions
