+++ src/TextRight.Core/Editing/DocumentEditorContext.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using JetBrains.Annotations;
 using TextRight.Core.Actions;
 using TextRight.Core.Cursors;
 using TextRight.Core.ObjectModel;
@@ -10,15 +9,6 @@ using TextRight.Core.ObjectModel.Cursors;
 
 namespace TextRight.Core
 {
-  public interface IDocumentEditorView : IDocumentItemView
-  {
-    /// <summary> Gets the block closest to the given point. </summary>
-    /// <param name="point"> The point at which the block should be queried. </param>
-    /// <returns> The closes block to the given point. </returns>
-    [NotNull]
-    Block GetBlockFor(DocumentPoint point);
-  }
-
   /// <summary> Represents a TextRight document that is being edited. </summary>
   public class DocumentEditorContext : EventEmitter, IDocumentItem
   {
