--- src/TextRight.Core/ObjectModel/IDocumentItemView.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using TextRight.Core.ObjectModel.Blocks;
-
-namespace TextRight.Core.ObjectModel
-{
-  /// <summary> A view item for a specific type of <see cref="IDocumentItem{T}"/>. </summary>
-  public interface IDocumentItemView : IEditorData
-  {
-    /// <summary> The DocumentItem associated with the View. </summary>
-    IDocumentItem DocumentItem { get; }
-  }
-}
