+++ src/TextRight.Editor/IDocumentItemView.cs
@@ -1,14 +1,16 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using TextRight.Core.ObjectModel;
 
 namespace TextRight.Editor
 {
   /// <summary>
   ///   A view for a document-item representation.
   /// </summary>
-  public interface IDocumentItemView
+  public interface IDocumentItemView : IEditorData
   {
-    
+    /// <summary> The DocumentItem associated with the View. </summary>
+    IDocumentItem DocumentItem { get; }
   }
 }
